require dependencies to be stable
As far as I can see no unstable dependencies are rquired to execute
the tests.

Also, keeping the composer.lock file does not add much value. It has
no affect when this library is required by another project as the
lock file is only taken into account for the root project anyway
while keeping it imposes the risk to forget to sync the composer.json
file with the lock file.